refact(404): abstract 404 link correction logic for collections to a component
